Test-TargetResource returning non-boolean value.
If only passing in the $Name parameter to Test-TargetResource, the only
required parameter, and $Name –eq $env:COMPUTERNAME, you drop to the
bottom of the function and the function completes without returning a
bool, causing an error.

Since all false cases are covered with existing return, a true return
was added to the end of the function.