DomainName check validates child without FQDN
The condition for verifying we're in the correct domain throws a non-terminating error if a child domain was used as input without FQDN. All other functions worked correctly in that scenario except validating the computer is in the correct domain. This change allows lacking FQDN to still return true value.