MSFT_xComputer:
Found and resolved bug with netbios being set to true and the computer name is returned as the domain name if not joined to the domain.

Changed output to return an empty string to match the [OutputType] and to work with unit tests.

Changed $Env:USERDOMAIN to use Get-Item instead so it can be mocked using Pester in the unit tests.

Changed Get-CimInstance -Class to -ClassName.

Written unit tests and updated MSFT_xComputer.Tests.ps1