ScheduledTask: Resolved inconsistent treatment of BuiltInAccount property. (#313)
* Resolved inconsistent treatment of BuiltInAccount property.

Added missing "NT Authority\" prefix to BuiltInAccount property in the resource Test method.
Added test cases using the ServiceAccount logonType and BuiltInAccount properties.

* Update DSC_ScheduledTask.Tests.ps1

Fixed unit test 'When a scheduled task is created using a Built In Service Account' - the Get-ScheduledTask mock needed updating to match the current observed behaviour.

* Update CHANGELOG.md

Moved changelog entry from the [Unreleased] to the Fixed heading. Linked to the issue fixed by this change.

* Update CHANGELOG.md

Moved line break in Changelog entry to make it clearer that the fix and issue are related.

* Update CHANGELOG.md

Moved WindowsCapability fix back to [Unreleased] at the request of PlagueHO

* Merge remote-tracking branch 'upstream/master'

* Quick fix up of CHANGELOG.MD